Use TESTS_TARGET_DIR envvar to override tests target dir
Rationale: At Debian, the autopkgtests are run inside a non-writable
directory, so the "target" subdirectory can't be used. Inside the
autopkgtest environment, we have an env variable AUTOPKGTEST_TMP
available which points to a temporary directory, that we can pass on to
the tests as TESTS_TARGET_DIR. This makes the tests pass properly inside
this environment.